Aplicacion de LL TP4 .v2
Faltaría solucionar un par de problemas con la función sort y pulir código

+++ TP4/Aplicacion LL TP4/estructura-afines.h
@@ -22,7 +22,7 @@ sJuego* videojuego_new();
  * \param sueldoStr int*
  * \return Employee*
  **/
-sJuego* videojuego_newParametros(int* idStr,char* nombreStr,int* horasJugablesStr,int* precioStr,int*stockStr);
+sJuego* videojuego_newParametros(int* idStr,char* nombreStr,int* horasJugablesStr,float* precioStr, int* stockStr);
 
 int videojuego_setStock(sJuego* juego,int stock);
